bcd: fix amalgamated build issues, const correctness
At least some distributions require _GNU_SOURCE to be set for stdio.h to
provide a prototype for asprintf(3). This introduces that definition and
reorders the Makefile.in (adding a comment) to ensure that macro is defined
before any includes are processed.

This also fixes various const-correctness issues found with -Wcast-qual
-Werror.