bcd: add bcd_associate_tid function
The API requires any thread that wishes to generate a trace to attach to the
socket. In systems with high thread counts, this can use many file
descriptors, while also adding additional latency to thread start time. This
new function allows changing the thread for a particular session, removing the
requirement for per-thread attaching (assuming structures are shared safely
between threads). This is expected to be useful in cases where threads may
be created and exit with frequency, and it is unknown which threads may be
trace candidates until they receive a signal.