Don't rely on the implicit order of IO events.
When a process potential tracee process exits all the ends of all the
pipes and sockets close, and the monitor process handles that in its
event loop.

The bcd_handler_sb handler exits the monitor on a hangup but we might
not handle that event before we handle the event of the end of the
pipe used for fatal hanging up.