Add support for abstract sockets on Linux.
If the unix socket path is specified starting with a '@' use an
abstract unix(7) socket instead of one on the filesystem.

This is only supported on Linux. It is also backwards compatible
because previously socket paths were required to be absolute paths,
so code that configured a path that didn't start with '/' would not
have worked anyway.

https://www.man7.org/linux/man-pages/man7/unix.7.html